AGeIDE Plotter
The AGeIDE now has a plotter tab! To activate the plot widget a button must be pressed first. This ensures that matplotlib does not slow down the program (after all, this widget won't be used most of the time.)

The AGeIDE.ConsoleWidget now has a new method called _setLocals which is used by setLocalsUpdateFunction as to allow simultaneous use of setLocalsUpdateFunction and setLocals. (Previously only one or the other could be used since setLocalsUpdateFunction would overwrite what ever was set via setLocals)