Help Window Update and IDE Help Text
The Help window is now created by the App to ensure that registering help pages works.
The Help window now searchers recursively for text. This way it is possible to link to subcategories.
AGeIDE now has some entries in the help window and pressing F1 brings up the help text for the current tab once the help page for the window has been brought up by pressing F1 once.