Help Window SubCat Overhaul
Overhauled the structure of the help window to allow for deeper nesting of the help tree.
Uniqueness of help entry names is now enforced on a branch level instead of a global level, allowing much more flexibility when naming help entries.
Changed the interface for opening help entries through the App to take a list of strings that represent the tree hierarchy (required due to uniqueness enforcement changes)