AGeIDE Method-Browser
Adds all builtins to the highlighting list for the CodeEditorWidget.
Adds a method browser for the OverloadWidget which allows selecting a method to edit via a modified InspectWidget.
Adds an "edit method" button to the inspect tab in the AGeIDE.
Overwriting methods of classes (opposed to methods of instances of classes) now properly saves the new code, allowing loading the code when wanting to apply further edits.