Merge TightGridWidget and TightGridFrame
TightGridWidget now inherits from QFrame instead of QWidget. The frame is not drawn by default.
TightGridFrame now inherits from TightGridWidget is is now merely a shorthand for TightGridWidget(..., frame=True).
This change removes quite some code duplication and makes it easier to add frames to WightGridWidget and widgets that inherit from it.
Especially, this makes it easy to add frames to the widgets from AGeInput.