Window-On-Top Fix For Linux
Using Alt+T to tell a window to always stay on top never worked properly on Linux. This never bothered me since Linux provides hotkeys for this. But I had an idea what caused the problem and I was right. So now it is finally fixed (though OS-hotkeys still work better)