Significantly improve the AGeIDE Autocomplete
Significantly improves the AGeIDE Autocomplete by actually parsing the active locals and globals.
The next step now is to support refreshing of the locals and globals to react to context changes. I might also consider parsing the editor itself for import statements to provide context for those, too, though that might become a bit messy.